move audio to a sepearte file
